nix: rename NixDerivation to NixExpression
This name better represents what this class does, which is to generate the
actual text of a Nix expression for a package.
